what if i move the mina explorer support here or closer
